Separated menu package
