+++ app/src/main/java/com/example/runtracker/MainActivity.kt
@@ -15,6 +15,8 @@ import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.runtracker.gallery.GalleryActivity
 import com.example.runtracker.history.HistoryActivity
+import com.example.runtracker.menu.MenuAdapter
+import com.example.runtracker.menu.MenuItem
 import com.example.runtracker.myAccount.MyAccountActivity
 import com.example.runtracker.runRecording.MapActivity
 import com.example.runtracker.settings.SettingsActivity
