+++ app/src/main/java/com/example/runtracker/menu/MenuItem.kt
@@ -1,3 +1,3 @@
-package com.example.runtracker
+package com.example.runtracker.menu
 
 class MenuItem(var data: String, var drawable: Int)
